pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lunes-cms"
description = "Content Management System for the Lunes App"
version = "2025.10.1"
authors = [{name = "Tür an Tür – Digitalfabrik gGmbH", email = "tech@integreat-app.de"}]
readme = "README.md"
requires-python = ">=3.11"
keywords = [
    "lunes",
    "cms",
    "django",
    "e-learning",
    "vocabulary",
    "trainer",
    "vocabulary-trainer",
    "visual-vocabulary-trainer",
]
license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE.md", "SECURITY.md"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Django :: 5.1",
    "Operating System :: POSIX :: Linux",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Natural Language :: English",
    "Natural Language :: German",
]
dependencies = [
	"django",
	"django-import-export",
	"django-jazzmin",
	"django-mptt",
	"django-qr-code",
	"djangorestframework",
	"drf-spectacular",
	"ipython",
	"openai",
	"pillow",
	"psycopg2",
	"pydub",
]

[project.optional-dependencies]
dev = [
	"black",
    "build",
	"bumpver",
	"pre-commit",
	"pyjwt",
	"pylint-django",
	"pylint-runner",
	"pytest-circleci-parallelized",
	"pytest-cov",
	"pytest-django",
	"pytest-icdiff",
	"pytest-xdist",
	"shellcheck-py",
	"sphinx",
	"sphinx-rtd-theme",
	"sphinx-last-updated-by-git",
	"sphinxcontrib-django",
	"twine",
    "wheel"
]
pinned = [
	"django==5.1.2",
	"django-import-export==4.3.13",
	"django-jazzmin==3.0.1",
	"django-mptt==0.18.0",
	"django-qr-code==4.2.0",
	"djangorestframework==3.16.1",
	"drf-spectacular==0.29.0",
	"ipython==9.6.0",
	"openai==2.6.1",
	"pillow==12.0.0",
	"psycopg2==2.9.11",
	"pydub==0.25.1",

]
dev-pinned = [
	"black==25.9.0",
    "build<=1.3.0",
	"bumpver==2025.1131",
	"pre-commit==4.3.0",
	"pyjwt==2.10.1",
	"pylint-django==2.6.1",
	"pylint-runner==0.7.0",
	"pytest-circleci-parallelized==0.1.0",
	"pytest-cov==7.0.0",
	"pytest-django==4.11.1",
	"pytest-icdiff==0.9",
	"pytest-xdist==3.8.0",
	"shellcheck-py==0.11.0.1",
	"sphinx==8.2.3",
	"sphinx-rtd-theme==3.0.2",
	"sphinx-last-updated-by-git==0.3.8",
	"sphinxcontrib-django==2.5",
	"twine==6.2.0",
	"wheel==0.45.1",
]

[project.urls]
"Homepage" = "https://lunes.app/"
"Source Code" = "https://github.com/digitalfabrik/lunes-cms"
"Issues" = "https://github.com/digitalfabrik/lunes-cms/issues"
"Release Notes" = "https://digitalfabrik.github.io/lunes-cms/release-notes.html"
"Documentation" = "https://digitalfabrik.github.io/lunes-cms/"

[tool.setuptools]
script-files = ["lunes_cms/lunes-cms-cli"]

[tool.setuptools.packages.find]
include = ["lunes_cms*"]

[tool.bumpver]
current_version = "2025.10.1"
version_pattern = "YYYY.MM.INC0[-TAG]"
commit_message = """
Bump version to {new_version}

[skip ci]
"""
commit = true
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"',
]
"lunes_cms/__init__.py" = [
    '__version__ = "{version}"',
]
"docs/src/conf.py" = [
    'release = "{version}"',
]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

